[refac] reorganize package structure to unify service definitions
