+++ edukit-core/src/main/java/com/edukit/core/common/service/EmailService.java
@@ -1,4 +1,4 @@
-package com.edukit.core.common.port;
+package com.edukit.core.common.service;
 
 public interface EmailService {
