+++ edukit-core/src/main/java/com/edukit/core/student/service/dto/StudentExcelRow.java
@@ -1,4 +1,4 @@
-package com.edukit.core.common.service.response;
+package com.edukit.core.student.service.dto;
 
 public record StudentExcelRow(
         String grade,
