+++ edukit-external/src/main/java/com/edukit/external/aws/s3/S3ServiceImpl.java
@@ -1,6 +1,6 @@
 package com.edukit.external.aws.s3;
 
-import com.edukit.core.common.port.FileStorageService;
+import com.edukit.core.common.service.FileStorageService;
 import com.edukit.core.common.service.response.UploadPresignedUrlResponse;
 import com.edukit.external.aws.s3.exception.S3ErrorCode;
 import com.edukit.external.aws.s3.exception.S3Exception;
